booking/src/main/java/com/dpw/runner/shipment/services/adapters/config/ShipmentServiceConfig.java
package com.dpw.runner.shipment.services.adapters.config;

import com.dpw.runner.shipment.services.aspects.MultitenancyAspect.RequestAuthContext;
import lombok.Data;
import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.http.HttpHeaders;
import org.springframework.http.MediaType;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestTemplate;

@Component
@Configuration
@ConfigurationProperties(prefix = "shipment")
@Data
public class ShipmentServiceConfig {
    private String baseUrl;
    private String createShipmentInV2Url;
    private String getIdByGuidUrl;
    private String createConsolidationFromBookingUrl;

    @Bean
    public RestTemplate restTemplateForShipment() {
        RestTemplate restTemplate = new RestTemplate();

        restTemplate.getInterceptors().add((request, body, execution) -> {
            HttpHeaders headers = request.getHeaders();
            headers.set("Authorization", RequestAuthContext.getAuthToken());
            headers.setContentType(MediaType.APPLICATION_JSON);
            return execution.execute(request, body);
        });
        return restTemplate;
    }
}
